UI gap fix: Split server communication and diagnostics tab.
