+++ frontend/src/app/components/server/server-communication-form/server-communication-form.html
@@ -1,0 +1,17 @@
+<!-- Copyright (C) 2016 NooBaa -->
+
+<section class="column">
+    <div class="column pad alt-bg push-next">
+        <h2 class="heading3">Internal Cluster Communication</h2>
+        <p>Testing communication between this server and other members of the cluster</p>
+    </div>
+    <div class="pad column">
+        <p class="push-next">Communication tests run automaticly every single minute</p>
+        <data-table class="communication-table alt-bg" params="
+            columns: columns,
+            data: otherServers,
+            rowFactory: server => createServerRow(server),
+            emptyMessage: 'No additional servers in the cluster'
+        "></data-table>
+    </div>
+</section>
