+++ frontend/src/app/components/server/server-communication-form/server-communication-form.less
@@ -1,0 +1,20 @@
+/* Copyright (C) 2016 NooBaa */
+
+server-communication-form {
+    display: block;
+
+    .communication-table {
+        .result-col {
+            min-width: 50px;
+        }
+
+        .name-col {
+            text-align: left;
+            width: 100%;
+        }
+
+        .address-col {
+            min-width: 100px;
+        }
+    }
+}
