+++ frontend/src/app/components/server/server-communication-form/test-result-row.js
