+++ frontend/src/app/components/server/server-diagnostics-form/server-diagnostics-form.less
@@ -2,20 +2,4 @@
 
 server-diagnostics-form {
     display: block;
-
-    .communication-table {
-        .result-col {
-            min-width: 50px;
-        }
-
-        .name-col {
-            text-align: left;
-            width: 100%;
-        }
-
-        .address-col {
-            min-width: 100px;
-        }
-
-    }
 }
